Plant Health Monitoring Models/requirements.txt.txt
opencv-python-headless
pillow
torch
torchvision
matplotlib
scikit-image
numpy
pandas